pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zhihu_md"
version = "0.0.1"
description = "把 markdown 转换为知乎能正常导入的 markdown。支持自动处理图片、公式"
authors = [
    {name = "DIYer22", email = "ylxx@live.com"},
]
maintainers = [
    {name = "DIYer22", email = "ylxx@live.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "chardet",
    "Pillow",
]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
keywords = ["zhihu_md"]

[project.scripts]
zhihu-md = "zhihu_md.main:main"

[project.urls]
Homepage = "https://github.com/DIYer22/zhihu_md"
Repository = "https://github.com/DIYer22/zhihu_md"
Issues = "https://github.com/DIYer22/zhihu_md/issues"

[tool.setuptools]
packages = ["zhihu_md"]
include-package-data = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=5.0",
    "mypy>=1.0",
]
